+++ src/extension.ts
@@ -1,6 +1,7 @@
 // The module 'vscode' contains the VS Code extensibility API
 // Import the module and reference it with the alias vscode in your code below
 import * as vscode from 'vscode';
+import { MarketplaceService } from './services/marketplaceService';
 
 // This method is called when your extension is activated
 // Your extension is activated the very first time the command is executed
@@ -13,10 +14,15 @@ export function activate(context: vscode.ExtensionContext) {
 	// The command has been defined in the package.json file
 	// Now provide the implementation of the command with registerCommand
 	// The commandId parameter must match the command field in package.json
-	const listMarketplace = vscode.commands.registerCommand('vscode-copilot-marketplace.listMarketplace', () => {
-		// The code you place here will be executed every time your command is executed
-		// Display a message box to the user
-		vscode.window.showInformationMessage('List of installed extensions in the marketplace.');
+	const listMarketplace = vscode.commands.registerCommand('vscode-copilot-marketplace.listMarketplace', async () => {
+		const service = new MarketplaceService();
+		const marketplaces = await service.getMarketplaces();
+
+		if (marketplaces.length === 0) {
+			vscode.window.showErrorMessage('No marketplaces found. Please add a marketplace repository first.');
+		} else {			
+			vscode.window.showQuickPick(marketplaces)
+		}
 	});
 
 	// Add a Marketplace command